service/llm_chat_app/requirements.txt
streamlit
neo4j
requests
python-dotenv
PyYAML
docker
temporalio
